refactor: split out pendingJobController methods so they are easier to reason
